requirements.txt
# Core dependencies
google-generativeai>=0.8.0
python-dotenv>=1.0.0
tenacity>=8.2.0
tqdm>=4.66.0

# AI Strategy Factory dependencies
pydantic>=2.0.0
perplexityai>=0.1.0

# Document generation
python-docx>=1.1.0
python-pptx>=0.6.23

# Web application
flask>=3.0.0
markdown>=3.5.0

# Utilities
httpx>=0.26.0
